Refactor tests and add specific failure test
